✅ test: small test adjustments

+++ src/unxt/_interop/unxt_interop_astropy/quantity.py
@@ -266,8 +266,8 @@ def ustrip(u: Any, x: AstropyQuantity) -> Any:
     >>> import unxt as u
 
     >>> x = apyu.Quantity(1000, "m")
-    >>> u.ustrip(u.unit("m"), x)
-    np.float64(1000.0)
+    >>> float(u.ustrip(u.unit("m"), x))
+    1000.0
 
     """
     return x.to_value(u)
